@@ -5,31 +5,45 @@
   "description": "frontend for npmcharts.com",
   "scripts": {
     "start": "npm run dev",
-    "dev": "gulp",
-    "build": "gulp build",
+    "dev": "webpack-dev-server --mode development --progress --hot",
+    "build": "webpack --mode production",
     "lint": "eslint .",
     "test": "jest"
   },
   "dependencies": {
     "array-to-sentence": "^1.1.0",
+    "babel-loader": "^8.0.4",
+    "copy-webpack-plugin": "4.6.0",
     "cors": "^2.8.4",
+    "css-loader": "1.0.1",
     "d3": "^3.5.6",
     "d3-shape": "^1.2.0",
     "date-fns": "^1.29.0",
+    "file-loader": "2.0.0",
     "filenamify": "^2.0.0",
     "filesize": "^3.6.0",
     "fs-extra": "^5.0.0",
-    "gulp": "~3.8.1",
-    "hammerjs": "^2.0.4",
+    "html-loader": "0.5.5",
+    "html-webpack-plugin": "^3.2.0",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "^4.17.4",
+    "mini-css-extract-plugin": "^0.4.4",
     "numeral": "^2.0.6",
     "nvd3": "1.8.3",
     "path-match": "^1.2.4",
     "path-to-regexp": "^2.3.0",
     "pm2": "^2.6.1",
-    "vue": "^1.0.21",
-    "vue-router": "^0.7.13",
+    "pug-loader": "^2.4.0",
+    "stylus": "0.54.5",
+    "stylus-loader": "^3.0.2",
+    "svg-inline-loader": "0.8.0",
+    "vue": "2",
+    "vue-router": "2",
+    "vue-template-compiler": "2.5.17",
+    "vue-template-loader": "1.0.0",
+    "webpack": "^4.25.1",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.10",
     "whatwg-fetch": "^0.9.0"
   },
   "devDependencies": {
@@ -45,29 +59,9 @@
     "eslint-plugin-jsx-a11y": "^5.0.3",
     "eslint-plugin-prettier": "^2.6.0",
     "eslint-plugin-react": "^7.6.1",
-    "exorcist": "^0.4.0",
-    "gulp": "~3.8.1",
-    "gulp-autoprefixer": "1.0.1",
-    "gulp-duration": "0.0.0",
-    "gulp-jade": "~0.9.0",
-    "gulp-replace": "^0.5.3",
-    "gulp-rev": "^4.0.0",
-    "gulp-sourcemaps": "^1.3.0",
-    "gulp-streamify": "0.0.5",
-    "gulp-stylus": "^2.6.0",
-    "gulp-uglify": "~1.0.1",
-    "gulp-util": "~3.0.1",
-    "gulp-watch": "^4.3.4",
-    "minimist": "^1.2.0",
-    "node-notifier": "^4.2.1",
     "npm-run-all": "^4.0.2",
     "prettier": "^1.10.2",
-    "regenerator-runtime": "^0.12.1",
-    "rimraf": "^2.3.4",
-    "stringify": "^3.1.0",
-    "vinyl-source-stream": "~1.0.0",
-    "vinyl-transform": "^1.0.0",
-    "watchify": "^3.2.1"
+    "regenerator-runtime": "^0.12.1"
   },
   "prettier": {
     "trailingComma": "all",
