@@ -1,12 +1,18 @@
-import Vue from 'vue';
+import _ from 'lodash';
 import { removePackage } from '../../packages/packages.js';
 import { isSameMonth, format as formatDate, addDays } from 'date-fns';
+import withRender from './legend.html';
 
-export default Vue.extend({
+export default withRender({
   props: {
     modules: Array,
     date: Date,
     periodLength: Number,
+  },
+  computed: {
+    sortedModules() {
+      return _.orderBy(this.modules, module => module.downloads, ['desc']);
+    },
   },
   template: require('./legend.html'),
   methods: {
